NEVISACCESSAPP-6290: added develop branch and workflow
- Changed app to be released for new architecture as that is the default for ReactNative as well.
- Aligned git tag creation with other repositories.
- Added develop and release workflows.
- Adjusted Fastfiles on both platforms.
- Added possibility to add private repositories during ci builds.
- Builds during prs are moved from gh workflow to Fastlane.
