+++ .github/actions/build-ios/action.yml
@@ -20,5 +20,9 @@ runs:
               new-arch-enabled: ${{ inputs.new-arch-enabled }}
 
         - name: Build iOS App
-          shell: bash
-          run: yarn react-native build-ios --mode Debug --verbose
+          env:
+            PRIVATE_NATIVE_REPOSITORY_URL: ${{ env.PRIVATE_NATIVE_IOS_REPOSITORY_URL }}
+          uses: maierj/fastlane-action@v3.1.0
+          with:
+            lane: 'pr'
+            subdirectory: 'ios'
