+++ .github/actions/setup-android/action.yml
@@ -15,6 +15,20 @@ runs:
               distribution: 'zulu'
               java-version: ${{ env.JAVA_VERSION }}
 
+        - name: Cache RubyGem Dependencies
+          uses: ruby/setup-ruby@v1
+          with:
+              ruby-version: ${{ env.RUBY_VERSION }}
+              bundler-cache: true
+              working-directory: 'android'
+
+        - name: Decode Keystore
+          uses: timheuer/base64-to-file@v1
+          with:
+              fileName: 'keystore-example-app.jks'
+              fileDir: './android/etc'
+              encodedString: ${{ env.KEYSTORE_FILE }}
+
         - name: Cache Gradle
           if: ${{ inputs.cache-gradle == 'true' }}
           uses: actions/cache@v4
