+++ .github/workflows/main.yml
@@ -18,7 +18,11 @@ on:
                     - "macos-15"
 
 env:
-    BUILD_TYPE: 'RELEASE' #BUILD_TYPE: Type of the build; allowed values: 'SNAPSHOT', 'RELEASE'
+    BUILD_TYPE: 'RELEASE' #BUILD_TYPE: Type of the build; allowed values: 'SNAPSHOT', 'RELEASE', 'RELEASE_CANDIDATE'
+    XCODE_VERSION: 16.2
+    JAVA_VERSION: 17
+    RUBY_VERSION: 3.1
+
     CURRENT_BUILD_URL: ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
     GH_USERNAME: ${{ vars.GH_USERNAME }}
     GH_PERSONAL_ACCESS_TOKEN: ${{ secrets.GH_PERSONAL_ACCESS_TOKEN }}
@@ -41,9 +45,6 @@ env:
     CONNECT_API_ISSUER_ID: ${{ secrets.CONNECT_API_ISSUER_ID }}
     CONNECT_API_KEY_CONTENT: ${{ secrets.CONNECT_API_KEY_CONTENT }}
     BETA_APP_REVIEW_INFO: ${{ vars.BETA_APP_REVIEW_INFO }}
-    XCODE_VERSION: 16.2
-    JAVA_VERSION: 17
-    RUBY_VERSION: 3.1
 
 jobs:
     lint-and-format:
@@ -95,10 +96,10 @@ jobs:
             - name: Checkout
               uses: actions/checkout@v4
 
-            - name: Release Android App with old arch
+            - name: Release Android App with new arch
               uses: ./.github/actions/release-android
               with:
-                  new-arch-enabled: false
+                  new-arch-enabled: true
                   version-number: ${{ needs.get-version-number.outputs.version-number }}
                   build-number: ${{ needs.get-version-number.outputs.build-number }}
 
@@ -109,10 +110,10 @@ jobs:
             - name: Checkout
               uses: actions/checkout@v4
 
-            - name: Release iOS App with old arch
+            - name: Release iOS App with new arch
               uses: ./.github/actions/release-ios
               with:
-                  new-arch-enabled: false
+                  new-arch-enabled: true
                   version-number: ${{ needs.get-version-number.outputs.version-number }}
                   build-number: ${{ needs.get-version-number.outputs.build-number }}
 
@@ -124,7 +125,8 @@ jobs:
               uses: actions/checkout@v4
 
             - name: Create git tag
-              uses: ./.github/actions/git-tag
+              uses: ./.github/actions/create-tag
               with:
-                  version-number: ${{ needs.get-version-number.outputs.version-number }}
+                  build-category: 'release'
+                  app-version: ${{ needs.get-version-number.outputs.version-number }}
                   build-number: ${{ needs.get-version-number.outputs.build-number }}
