+++ android/fastlane/.env.example
@@ -1,0 +1,6 @@
+CURRENT_BUILD_URL=
+FIREBASE_APP_ID_ANDROID=
+PRIVATE_NATIVE_REPOSITORY_URL=
+SIGNING_CONFIGS=
+SLACK_WEBHOOK_URL=
+NEW_ARCH_ENABLED=
