+++ android/fastlane/Fastfile
@@ -7,8 +7,10 @@ default_platform(:android)
 
 root_dir = File.expand_path(File.join(File.dirname(__FILE__), "../"))
 signing_configs_gradle = File.join(root_dir, "app/signing_configs.gradle")
-apk_path = File.join(root_dir, "../build/app/outputs/flutter-apk/app-debug.apk")
 display_name = "Nevis Mobile Authentication SDK Example App React Android"
+app_build_gradle = File.join(root_dir, "app/build.gradle")
+build_gradle = File.join(root_dir, "build.gradle")
+public_maven_repository_url = "https://maven.pkg.github.com/nevissecurity/nevis-mobile-authentication-sdk-android-package"
 
 def report_on_success(message)
 	slack(
@@ -43,16 +45,63 @@ def get_boolean?(input)
 end
 
 platform :android do
-	desc "Build and distribute the application"
+	desc "Pull request build"
+	desc ""
+	lane :pr do
+		utilize_private_repos
+		build
+	end
+
+	desc "Continuous build"
 	desc "#### Options"
 	desc " * **`version`**: The version of the application."
 	desc " * **`build_number`**: The build number of the application."
-	desc " * **`new_arch_enabled`**: The build number of the application."
+	desc ""
+	lane :develop do |options|
+		utilize_private_repos
+		distribute(options)
+	end
+
+	desc "Release build"
+	desc "#### Options"
+	desc " * **`version`**: The version of the application."
+	desc " * **`build_number`**: The build number of the application."
 	desc ""
 	lane :main do |options|
+		distribute(options)
+	end
+
+	####
+	#### Private lanes
+	####
+
+	desc "Utilize private repositories"
+	desc ""
+	private_lane :utilize_private_repos do
+		private_maven_repository_url = ENV["PRIVATE_NATIVE_REPOSITORY_URL"]
+		modify_file(
+			file_path: build_gradle,
+			old_value: "url \"#{public_maven_repository_url}\"",
+			new_value: "url \"#{private_maven_repository_url}\"",
+		) if private_maven_repository_url
+	end
+
+	desc "Build the application"
+	desc ""
+	private_lane :build do
+		new_arch_enabled = ENV["NEW_ARCH_ENABLED"].nil? ? true : get_boolean?(ENV["NEW_ARCH_ENABLED"])
+		sh("yarn react-native build-android --mode Debug --extra-params -PnewArchEnabled=#{new_arch_enabled}")
+	end
+
+	desc "Distribute the application"
+	desc "#### Options"
+	desc " * **`version`**: The version of the application."
+	desc " * **`build_number`**: The build number of the application."
+	desc ""
+	private_lane :distribute do |options|
 		version = options[:version]
 		build_number = options[:build_number]
-		new_arch_enabled = options[:new_arch_enabled].nil? ? false : get_boolean?(options[:new_arch_enabled])
+		new_arch_enabled = ENV["NEW_ARCH_ENABLED"].nil? ? true : get_boolean?(ENV["NEW_ARCH_ENABLED"])
 		UI.message("Distributing #{display_name} #{version} (#{build_number}) 📦")
 
 		modify_gradle_file(
@@ -67,15 +116,17 @@ platform :android do
 
 		sh("echo \"#{ENV['SIGNING_CONFIGS']}\" > #{signing_configs_gradle}")
 
-		modify_gradle_file(
-			constant: "signingConfig",
-			value: "signingConfigs.signing",
+		modify_file(
+			file_path: app_build_gradle,
+			old_value: "signingConfig signingConfigs.debug",
+			new_value: "signingConfig signingConfigs.signing",
 		)
 
-		modify_gradle_file(
-			constant: "apply plugin: 'com.facebook.react'",
-			value: "apply from: 'signing_configs.gradle'",
-			mode: "append"
+		modify_file(
+			file_path: app_build_gradle,
+			old_value: "android {",
+			new_value: "apply from: 'signing_configs.gradle'\n",
+			mode: "prepend"
 		)
 
 		gradle(
