+++ ios/fastlane/.env.example
@@ -1,0 +1,13 @@
+BETA_APP_REVIEW_INFO=
+CONNECT_API_ISSUER_ID=
+CONNECT_API_KEY_CONTENT=
+CONNECT_API_KEY_ID=
+CODE_SIGNING_IDENTITY=
+CURRENT_BUILD_URL=
+DEVELOPER_PORTAL_TEAM_ID=
+FIREBASE_APP_ID_IOS=
+MATCH_GIT_URL=
+PRIVATE_NATIVE_REPOSITORY_URL=
+PROVISIONING_PROFILE_SPECIFIER=
+SLACK_WEBHOOK_URL=
+TEMP_KEYCHAIN_NAME=
