+++ ios/fastlane/Fastfile
@@ -57,13 +57,72 @@ def report_on_error(message, exception)
 	UI.user_error!(exception.to_s)
 end
 
+def prepare_podfile(private_pod_repository_url)
+	filename = File.join("./../Podfile")
+	File.open(filename, "r") do |original_lines|
+		File.unlink(filename)
+		File.open(filename, "w") do |new_lines|
+			new_lines.write "source '#{private_pod_repository_url}'\n"
+			new_lines.write(original_lines.read)
+		end
+	end
+end
+
 platform :ios do
-	desc "Build and distribute the application"
+	desc "Pull request build"
+	desc ""
+	lane :pr do
+		utilize_private_repos
+		build
+	end
+
+	desc "Continuous build"
+	desc "#### Options"
+	desc " * **`version`**: The version of the application."
+	desc " * **`build_number`**: The build number of the application."
+	desc ""
+	lane :develop do |options|
+		utilize_private_repos
+		distribute(options)
+	end
+
+	desc "Release build"
 	desc "#### Options"
 	desc " * **`version`**: The version of the application."
 	desc " * **`build_number`**: The build number of the application."
 	desc ""
 	lane :main do |options|
+		distribute(options)
+	end
+
+	####
+	#### Private lanes
+	####
+
+	desc "Utilize private repositories"
+	desc ""
+	private_lane :utilize_private_repos do
+		private_pod_repository_url = ENV["PRIVATE_NATIVE_REPOSITORY_URL"]
+		prepare_podfile(private_pod_repository_url) if private_pod_repository_url
+	end
+
+	desc "Build the application"
+	desc ""
+	private_lane :build do
+		UI.message("Build Debug version")
+		Dir.chdir("..") do
+			Actions.sh("yarn react-native build-ios --mode Debug", error_callback: ->(result) {
+				raise "❌ Build Debug version failed failed! Result: #{result}"
+			})
+		end
+	end
+
+	desc "Distribute the application"
+	desc "#### Options"
+	desc " * **`version`**: The version of the application."
+	desc " * **`build_number`**: The build number of the application."
+	desc ""
+	private_lane :distribute do |options|
 		version = options[:version]
 		build_number = options[:build_number]
 
@@ -83,8 +142,8 @@ platform :ios do
 			build_number:
 		)
 
-		distribute(destination: Destination::FIREBASE)
-		distribute(destination: Destination::TESTFLIGHT)
+		upload(destination: Destination::FIREBASE)
+		upload(destination: Destination::TESTFLIGHT)
 
 		delete_temp_keychain(keychain_name)
 		report_on_success("#{display_name} build completed: #{version} (#{build_number}) ✅")
@@ -93,7 +152,11 @@ platform :ios do
 		report_on_error("#{display_name} build failed ❌", e)
 	end
 
-	private_lane :distribute do |options|
+	desc "Upload the application to the specified destination"
+	desc "#### Options"
+	desc " * **`destination`**: The destination to upload the application to."
+	desc ""
+	private_lane :upload do |options|
 		destination = options[:destination]
 		keychain_name = ENV["TEMP_KEYCHAIN_NAME"]
 
@@ -156,23 +219,9 @@ platform :ios do
 		end
 	end
 
-	##
-	## Utilities
-	##
-
-	module Step
-		MATCH = "match".freeze
-		CODE_SIGN = "code_sign".freeze
-		GYM = "gym".freeze
-		PROVISIONING_PROFILE = "prov_profile".freeze
-	end
-
-	module Destination
-		FIREBASE = "firebase".freeze
-		TESTFLIGHT = "testflight".freeze
-	end
-
-	private_lane :testflight_app_distribution do |_options|
+	desc "Upload the application to TestFlight"
+	desc ""
+	private_lane :testflight_app_distribution do
 		# Get AppStoreConnect Credentials
 		app_store_connect_api_key(
 			key_id: ENV["CONNECT_API_KEY_ID"],
@@ -200,6 +249,27 @@ platform :ios do
 		)
 	end
 
+	####
+	#### Utilities
+	####
+
+	module Step
+		MATCH = "match".freeze
+		CODE_SIGN = "code_sign".freeze
+		GYM = "gym".freeze
+		PROVISIONING_PROFILE = "prov_profile".freeze
+	end
+
+	module Destination
+		FIREBASE = "firebase".freeze
+		TESTFLIGHT = "testflight".freeze
+	end
+
+	desc "Get configuration value"
+	desc "#### Options"
+	desc " * **`destination`**: The destination to upload the application to. Possible values, see Destination module."
+	desc " * **`step`**: The step to get the configuration value for. Possible values, see Step module."
+	desc ""
 	private_lane :getConfig do |options|
 		UI.user_error!("⚠️ destination is not provided") unless options[:destination]
 		UI.user_error!("⚠️ step is not provided") unless options[:step]
